moved all open dbus function variants into a single header and source file
