--- src/libhirte/bus/user-bus.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <systemd/sd-bus.h>
-
-sd_bus *user_bus_open(sd_event *event);
