+++ src/libhirte/meson.build
@@ -1,12 +1,6 @@
 libhirte_src = [
-    'bus/peer-bus.c',
-    'bus/peer-bus.h',
-    'bus/systemd-bus.c',
-    'bus/systemd-bus.h',
-    'bus/user-bus.c',
-    'bus/user-bus.h',
-    'bus/system-bus.c',
-    'bus/system-bus.h',
+    'bus/bus.c',
+    'bus/bus.h',
     'bus/utils.c',
     'bus/utils.h',
     'common/common.h',
